+++ frontend/src/components/search/filter/SearchAndFilterOptionsCard.vue
